member initilizer work.

+++ modernCPP/courseWork/06-memberInitializer/Vector3f.cpp
@@ -1,5 +1,5 @@
 #include "Vector3f.hpp"
 
-Vector3f::Vector3f(int inX, int inY, int inZ) 
+Vector3f::Vector3f(float inX, float inY, float inZ) 
             : x(inX), y(inY), z(inZ) {
 }
